feat: hash resources to track modifications
If you run a tool to modify the generated resource files (e.g. prettier
or eslint) before checking in the generated model files there is no way
for Anchor to know whether or not it must write to a file again.

This means that all generated files will be written to _and_
prettier/whatever codemod tool will have to be run on all the generated
files each time a change is made. This can take a while if you have a
lot of resources.

To reduce time spent on codemod we hash the file content _before_
writing it to the file and save that. Then, on the next generation,
we'll check that file and won't write to it. The modified files are
returned so they can be used to run the codemods on relevant files.

```sh
cd spec/example
rails anchor:multigen | xargs prettier --write
```